Build user registration and login
These changes add a UserController for implementing user registration
and a SessionController for providing log in and log out functionality.

The salt field of the User schema is removed since the Pbkdf2 package
stores the salt and hash as a single encoded string. The entire encoded
string is stored in the password_hash field.