Build workspace controller and views
This adds views for creating workspaces and presenting a list of
channels within a workspace. Both require user authentication. The
channel list can only be accessed by users who are members of the
workspace. Creating a workspace performs a transaction that also adds
the user as a member and admin of the workspace.

Completes #2.